requirements.txt
aniso8601==9.0.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-MySQL==1.6.0
Flask-RESTful==0.3.10
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
